moving viewmodel to new arch
