+++ src/adapters/controllers/confirm_change_password_controller.py
@@ -1,9 +1,7 @@
-from pydantic import ValidationError
-
 from src.adapters.errors.http_exception import HttpException
 from src.adapters.helpers.http_models import BadRequest, HttpRequest, HttpResponse, InternalServerError, Ok
-from src.adapters.viewmodels.change_password_model import ChangePasswordModel
-from src.adapters.viewmodels.login_user_model import LoginUserModel
+from src.modules.change_password.app.change_password_viewmodel import ChangePasswordModel
+from src.modules.login_user.app.login_user_viewmodel import LoginUserModel
 from src.domain.errors.errors import UnexpectedError, EntityError, NonExistentUser, InvalidCredentials
 from src.domain.repositories.user_repository_interface import IUserRepository
 from src.domain.usecases.change_password_usecase import ChangePasswordUsecase
@@ -13,7 +11,8 @@ from src.domain.usecases.login_user_usecase import LoginUserUsecase
 
 class ConfirmChangePasswordController:
     def __init__(self, userRepository: IUserRepository) -> None:
-        self._confirmChangePasswordUsecase = ConfirmChangePasswordUsecase(userRepository)
+        self._confirmChangePasswordUsecase = ConfirmChangePasswordUsecase(
+            userRepository)
 
     async def __call__(self, req: HttpRequest) -> HttpResponse:
