+++ src/adapters/controllers/confirm_user_creation_controller.py
@@ -1,10 +1,8 @@
-from pydantic import ValidationError
-
 from src.adapters.errors.http_exception import HttpException
 from src.adapters.helpers.http_models import BadRequest, HttpRequest, HttpResponse, InternalServerError, Ok, \
     RedirectResponse, NotFound
-from src.adapters.viewmodels.change_password_model import ChangePasswordModel
-from src.adapters.viewmodels.login_user_model import LoginUserModel
+from src.modules.change_password.app.change_password_viewmodel import ChangePasswordModel
+from src.modules.login_user.app.login_user_viewmodel import LoginUserModel
 from src.domain.errors.errors import UnexpectedError, EntityError, NonExistentUser, InvalidCredentials, \
     UserAlreadyConfirmed
 from src.domain.repositories.user_repository_interface import IUserRepository
@@ -16,7 +14,8 @@ from src.domain.usecases.login_user_usecase import LoginUserUsecase
 
 class ConfirmUserCreationController:
     def __init__(self, userRepository: IUserRepository) -> None:
-        self._confirmUserCreationUsecase = ConfirmUserCreationUsecase(userRepository)
+        self._confirmUserCreationUsecase = ConfirmUserCreationUsecase(
+            userRepository)
 
     async def __call__(self, req: HttpRequest) -> HttpResponse:
