+++ src/adapters/controllers/get_user_by_cpfrne_controller.py
@@ -1,6 +1,6 @@
 from src.adapters.errors.http_exception import HttpException
 from src.adapters.helpers.http_models import HttpRequest, HttpResponse, BadRequest, Ok, NoContent, InternalServerError
-from src.adapters.viewmodels.get_user_model import GetUserModel
+from src.modules.get_user.app.get_user_viewmodel import GetUserModel
 from src.domain.entities.user import User
 from src.domain.errors.errors import UnexpectedError, NoItemsFound, NonExistentUser
 from src.domain.usecases.get_user_by_cpfrne_usecase import GetUserByCpfRneUsecase
@@ -10,7 +10,8 @@ from src.domain.repositories.user_repository_interface import IUserRepository
 class GetUserByCpfRneController:
 
     def __init__(self, userRepository: IUserRepository) -> None:
-        self._getAllUserByCpfRneUseCase = GetUserByCpfRneUsecase(userRepository)
+        self._getAllUserByCpfRneUseCase = GetUserByCpfRneUsecase(
+            userRepository)
 
     async def __call__(self, req: HttpRequest) -> HttpResponse:
 
@@ -18,7 +19,7 @@ class GetUserByCpfRneController:
             return BadRequest('Missing parameter.')
 
         try:
-            if  type(req.query['cpfRne']) != int:
+            if type(req.query['cpfRne']) != int:
                 return BadRequest('Invalid parameter. (Cpf value should be Int) ')
 
             user = await self._getAllUserByCpfRneUseCase(int(req.query['cpfRne']))
@@ -38,6 +39,3 @@ class GetUserByCpfRneController:
         except Exception as e:
             err = InternalServerError(e.args[0])
             return err
-
-
-
