+++ src/adapters/controllers/resend_creation_confirmation_controller.py
@@ -1,10 +1,8 @@
-from pydantic import ValidationError
-
 from src.adapters.errors.http_exception import HttpException
 from src.adapters.helpers.http_models import BadRequest, HttpRequest, HttpResponse, InternalServerError, Ok, \
     RedirectResponse, NotFound
-from src.adapters.viewmodels.change_password_model import ChangePasswordModel
-from src.adapters.viewmodels.login_user_model import LoginUserModel
+from src.modules.change_password.app.change_password_viewmodel import ChangePasswordModel
+from src.modules.login_user.app.login_user_viewmodel import LoginUserModel
 from src.domain.errors.errors import UnexpectedError, EntityError, NonExistentUser, InvalidCredentials, \
     UserAlreadyConfirmed
 from src.domain.repositories.user_repository_interface import IUserRepository
@@ -17,7 +15,8 @@ from src.domain.usecases.resend_creation_confirmation_usecase import ResendCreat
 
 class ResendCreationConfirmationController:
     def __init__(self, userRepository: IUserRepository) -> None:
-        self._resendCreationConfirmationUsecase = ResendCreationConfirmationUsecase(userRepository)
+        self._resendCreationConfirmationUsecase = ResendCreationConfirmationUsecase(
+            userRepository)
 
     async def __call__(self, req: HttpRequest) -> HttpResponse:
 
@@ -28,7 +27,8 @@ class ResendCreationConfirmationController:
 
         try:
 
-            req.body['cpfRne'] = req.body['cpf_rne'].replace('.', '').replace('-', '').replace(' ', '')
+            req.body['cpfRne'] = req.body['cpf_rne'].replace(
+                '.', '').replace('-', '').replace(' ', '')
 
             result = await self._resendCreationConfirmationUsecase(cpfRne=str(req.body['cpfRne']))
